Explosions penetrate players
