+++ src/game/shared/tf/tf_weapon_wrench.h
@@ -16,24 +16,6 @@
 #define CTFWrench C_TFWrench
 #endif
 
-class CTraceFilterIgnorePlayers : public CTraceFilterSimple
-{
-public:
-	// It does have a base, but we'll never network anything below here..
-	DECLARE_CLASS( CTraceFilterIgnorePlayers, CTraceFilterSimple );
-
-	CTraceFilterIgnorePlayers( const IHandleEntity *passentity, int collisionGroup )
-		: CTraceFilterSimple( passentity, collisionGroup )
-	{
-	}
-
-	virtual bool ShouldHitEntity( IHandleEntity *pServerEntity, int contentsMask )
-	{
-		CBaseEntity *pEntity = EntityFromEntityHandle( pServerEntity );
-		return pEntity && !pEntity->IsPlayer();
-	}
-};
-
 //=============================================================================
 //
 // Wrench class.
